+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Next release
+* Add thermostat support
 * Add camera support
 * Add low speed support for roller shutters
 * roller shutter position is no more a property