+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Next release
+*
+### O.3.0
+* Add thermostat support
 * Add camera support
 * Add low speed support for roller shutters
 * roller shutter position is no more a property